requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.1
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
motor>=3.0.0
email-validator>=2.0.0
python-dotenv>=1.0.0
torch>=2.0.0
transformers>=4.30.0 